router: Refactor and add export for Remix
Adds a dedicated export for use in Remix, and refactors to support this
change. In the future we should consider supporting Next.js in the same
way.
