+++ lib/router/spec/is-active.ts
@@ -1,118 +1,123 @@
-import { useIsActive } from '../src/is-active';
+import { makeUseIsActive } from '../src/is-active';
 
-jest.mock('../src/router', () => ({
-  useLocation() {
-    const { hash, pathname, search, searchParams } = new URL('https://user:pass@localhost:8080/path/to/resource?foo=bar#baz');
+const useLocation = () => {
+  const { hash, pathname, search, searchParams } = new URL('https://user:pass@localhost:8080/path/to/resource?foo=bar#baz');
 
-    return {
-      hash,
-      key: '',
-      pathname,
-      query: Object.fromEntries(searchParams.entries()),
-      search,
-      searchParams,
-      state: undefined
-    };
-  }
-}));
+  return {
+    hash,
+    key: '',
+    pathname,
+    query: Object.fromEntries(searchParams.entries()),
+    search,
+    searchParams,
+    state: undefined
+  };
+};
 
-describe('useIsActive', () => {
-  it('is a function', () => expect(useIsActive).toBeInstanceOf(Function));
-  it('that takes no parameters', () => expect(useIsActive).toHaveLength(0));
+describe('makeUseIsActive', () => {
+  it('is a function', () => expect(makeUseIsActive).toBeInstanceOf(Function));
+  it('that takes one parameters', () => expect(makeUseIsActive).toHaveLength(1));
 
-  describe('when called', () => {
-    const isActive = useIsActive();
+  describe('when called / useIsActive', () => {
+    const useIsActive = makeUseIsActive(useLocation);
 
-    it('returns a function', () => expect(isActive).toBeInstanceOf(Function));
-    it('that takes one parameter', () => expect(isActive).toHaveLength(1));
+    it('is a function', () => expect(useIsActive).toBeInstanceOf(Function));
+    it('that takes no parameters', () => expect(useIsActive).toHaveLength(0));
 
-    describe('the returned function', () => {
-      describe('when given the current href', () => {
-        const href = '/path/to/resource?foo=bar#baz';
-        const result = isActive(href);
+    describe('when called', () => {
+      const isActive = useIsActive();
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns true', () => expect(result).toEqual(true));
-      });
+      it('returns a function', () => expect(isActive).toBeInstanceOf(Function));
+      it('that takes one parameter', () => expect(isActive).toHaveLength(1));
 
-      describe('when given the current href without hash', () => {
-        const href = '/path/to/resource?foo=bar';
-        const result = isActive(href);
+      describe('the returned function', () => {
+        describe('when given the current href', () => {
+          const href = '/path/to/resource?foo=bar#baz';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns true', () => expect(result).toEqual(true));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns true', () => expect(result).toEqual(true));
+        });
 
-      describe('when given the current href without query string', () => {
-        const href = '/path/to/resource#baz';
-        const result = isActive(href);
+        describe('when given the current href without hash', () => {
+          const href = '/path/to/resource?foo=bar';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns true', () => expect(result).toEqual(true));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns true', () => expect(result).toEqual(true));
+        });
 
-      describe('when given a different href', () => {
-        const href = '/path/to/other?foo=bar#baz';
-        const result = isActive(href);
+        describe('when given the current href without query string', () => {
+          const href = '/path/to/resource#baz';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns false', () => expect(result).toEqual(false));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns true', () => expect(result).toEqual(true));
+        });
 
-      describe('when given a relative href to the same resource', () => {
-        const href = '../to/resource?foo=bar#baz';
-        const result = isActive(href);
+        describe('when given a different href', () => {
+          const href = '/path/to/other?foo=bar#baz';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns true', () => expect(result).toEqual(true));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns false', () => expect(result).toEqual(false));
+        });
 
-      describe('when given a relative href to a different resource', () => {
-        const href = '../to/other?foo=bar#baz';
-        const result = isActive(href);
+        describe('when given a relative href to the same resource', () => {
+          const href = '../to/resource?foo=bar#baz';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns false', () => expect(result).toEqual(false));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns true', () => expect(result).toEqual(true));
+        });
 
-      describe('when given \'/\' as a href', () => {
-        const href = '/';
-        const result = isActive(href);
+        describe('when given a relative href to a different resource', () => {
+          const href = '../to/other?foo=bar#baz';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns false', () => expect(result).toEqual(false));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns false', () => expect(result).toEqual(false));
+        });
 
-      describe('when given an empty href', () => {
-        const href = '';
-        const result = isActive(href);
+        describe('when given \'/\' as a href', () => {
+          const href = '/';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns true', () => expect(result).toEqual(true));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns false', () => expect(result).toEqual(false));
+        });
 
-      describe('when given just the query string', () => {
-        const href = '?foo=bar';
-        const result = isActive(href);
+        describe('when given an empty href', () => {
+          const href = '';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns true', () => expect(result).toEqual(true));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns true', () => expect(result).toEqual(true));
+        });
 
-      describe('when given just a new query string', () => {
-        const href = '?foo=baz';
-        const result = isActive(href);
+        describe('when given just the query string', () => {
+          const href = '?foo=bar';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns false', () => expect(result).toEqual(false));
-      });
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns true', () => expect(result).toEqual(true));
+        });
 
-      describe('when given a hash link', () => {
-        const href = '#hash';
-        const result = isActive(href);
+        describe('when given just a new query string', () => {
+          const href = '?foo=baz';
+          const result = isActive(href);
 
-        it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
-        it('returns true', () => expect(result).toEqual(true));
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns false', () => expect(result).toEqual(false));
+        });
+
+        describe('when given a hash link', () => {
+          const href = '#hash';
+          const result = isActive(href);
+
+          it('returns a boolean', () => expect(typeof result).toEqual('boolean'));
+          it('returns true', () => expect(result).toEqual(true));
+        });
       });
     });
   });
